start splitting Mimer up in more modules
