+++ src/full/Agda/Mimer/Mimer.hs
@@ -27,44 +27,36 @@ module Agda.Mimer.Mimer
 
 import Prelude hiding (null)
 
-import Control.DeepSeq (force, NFData(..))
+import Control.DeepSeq (NFData(..))
 import Control.Monad
 import Control.Monad.Except (catchError)
 import Control.Monad.Error.Class (MonadError)
 import Control.Monad.IO.Class (liftIO)
 import Control.Monad.Reader (ReaderT(..), runReaderT, asks, ask, lift)
-import Data.Function (on)
 import Data.Functor ((<&>))
-import Data.List (sortOn, intersect, transpose, (\\))
+import Data.List (sortOn, intersect, (\\))
 import qualified Data.List.NonEmpty as NonEmptyList (head)
-import Data.Map (Map)
 import qualified Data.Map as Map
 import Data.Set (Set)
 import qualified Data.Set as Set
-import Data.Maybe (maybeToList, fromMaybe, maybe, isNothing)
+import Data.Maybe (maybeToList, fromMaybe, isNothing)
 import Data.PQueue.Min (MinQueue)
 import qualified Data.PQueue.Min as Q
-import GHC.Generics (Generic)
-import qualified Text.PrettyPrint.Boxes as Box
-import qualified Data.Text as Text
 
 import qualified Agda.Benchmarking as Bench
-import Agda.Interaction.MakeCase (makeCase, getClauseZipperForIP, recheckAbstractClause)
-import Agda.Syntax.Abstract (Expr(AbsurdLam))
+import Agda.Syntax.Abstract (Expr)
 import qualified Agda.Syntax.Abstract as A
 import qualified Agda.Syntax.Abstract.Views as A
-import Agda.Syntax.Abstract.Name (QName(..), Name(..))
-import Agda.Syntax.Common (InteractionId(..), MetaId(..), ArgInfo(..), defaultArgInfo, Origin(..), ConOrigin(..), Hiding(..), setOrigin, NameId, Nat, namedThing, Arg(..), setHiding, getHiding, ProjOrigin(..), rangedThing, woThing, nameOf, visible)
+import Agda.Syntax.Common
 import Agda.Syntax.Common.Pretty (Pretty)
 import qualified Agda.Syntax.Common.Pretty as P
 import qualified Agda.Syntax.Concrete.Name as C
-import Agda.Syntax.Info (pattern UnificationMeta, exprNoRange)
+import Agda.Syntax.Info (pattern UnificationMeta)
 import Agda.Syntax.Internal
 import Agda.Syntax.Internal.MetaVars (AllMetas(..))
-import Agda.Syntax.Internal.Pattern (clausePerm)
 import Agda.Syntax.Position (Range, rangeFile, rangeFilePath, noRange)
 import qualified Agda.Syntax.Scope.Base as Scope
-import Agda.Syntax.Translation.InternalToAbstract (reify, NamedClause(..), blankNotInScope)
+import Agda.Syntax.Translation.InternalToAbstract (reify, blankNotInScope)
 import Agda.Syntax.Translation.AbstractToConcrete (abstractToConcrete_)
 
 import Agda.TypeChecking.Primitive (getBuiltinName)
@@ -72,17 +64,15 @@ import Agda.TypeChecking.Constraints (noConstraints)
 import Agda.TypeChecking.Datatypes (isDataOrRecord)
 import Agda.TypeChecking.Conversion (equalType)
 import Agda.TypeChecking.Empty (isEmptyType)
-import Agda.TypeChecking.Free (flexRigOccurrenceIn, freeVars)
 import Agda.TypeChecking.Level (levelType)
 import Agda.TypeChecking.MetaVars (newValueMeta)
 import Agda.TypeChecking.Monad -- (MonadTCM, lookupInteractionId, getConstInfo, liftTCM, clScope, getMetaInfo, lookupMeta, MetaVariable(..), metaType, typeOfConst, getMetaType, MetaInfo(..), getMetaTypeInContext)
 import Agda.TypeChecking.Pretty
-import Agda.TypeChecking.Records (isRecord, isRecursiveRecord)
+import Agda.TypeChecking.Records (isRecord)
 import Agda.TypeChecking.Reduce (reduce, instantiateFull, instantiate)
-import Agda.TypeChecking.Rules.LHS.Problem (AsBinding(..))
 import Agda.TypeChecking.Rules.Term  (makeAbsurdLambda)
 import Agda.TypeChecking.Substitute (apply, applyE, piApply, NoSubst(..), pattern TelV, telView')
-import Agda.TypeChecking.Telescope (piApplyM, flattenTel, teleArgs)
+import Agda.TypeChecking.Telescope (piApplyM, flattenTel)
 
 import Agda.Utils.Benchmark (billTo)
 import Agda.Utils.FileName (filePath)
@@ -95,12 +85,8 @@ import qualified Agda.Utils.Maybe.Strict as SMaybe
 import Agda.Utils.Null
 import Agda.Utils.Time (CPUTime(..), getCPUTime, fromMilliseconds)
 import Agda.Utils.Tuple (mapFst, mapSnd)
-import Agda.Utils.FileName (AbsolutePath(..))
 
-import Agda.Mimer.Options
-
-import System.IO.Unsafe (unsafePerformIO)
-import Data.IORef (IORef, writeIORef, readIORef, newIORef, modifyIORef')
+import Data.IORef (readIORef, newIORef)
 
 -- Temporary (used for custom cost verbosity hack)
 import qualified Agda.Utils.Maybe.Strict as Strict
@@ -110,14 +96,9 @@ import Agda.Interaction.BasicOps (normalForm)
 import Agda.Interaction.Options.Base (parseVerboseKey)
 import Agda.Utils.List (lastWithDefault)
 
-data MimerResult
-  = MimerExpr String -- ^ Returns 'String' rather than 'Expr' because the give action expects a string.
-  | MimerClauses QName [A.Clause]
-  | MimerList [(Int, String)]
-  | MimerNoResult
-  deriving (Generic)
-
-instance NFData MimerResult
+import Agda.Mimer.Types
+import Agda.Mimer.Monad
+import Agda.Mimer.Options
 
 -- | Entry point.
 --   Run Mimer on the given interaction point, returning the desired solution(s).
@@ -176,148 +157,6 @@ mimer norm ii rng argStr = liftTCM $ do
 ------------------------------------------------------------------------------
 -- * Data types
 ------------------------------------------------------------------------------
-
-type SM a = ReaderT SearchOptions TCM a
-
-data SearchBranch = SearchBranch
-  { sbTCState :: TCState
-  , sbGoals :: [Goal]
-  , sbCost :: Int
-  , sbCache :: Map CheckpointId ComponentCache
-  , sbComponentsUsed :: Map Name Int -- ^ Number of times each component has been used
-  }
-  deriving (Generic)
-instance NFData SearchBranch
-
--- | NOTE: Equality is only on the fields `sbCost` and `sbGoals`
-instance Eq SearchBranch where
-  sb1 == sb2 = sbCost sb1 == sbCost sb2 && sbGoals sb1 == sbGoals sb2
-
--- TODO: Explain
-instance Ord SearchBranch where
-  compare = compare `on` sbCost
-
--- Map source component to generated components
-type ComponentCache = Map Component (Maybe [Component])
-
-data Goal = Goal
-  { goalMeta :: MetaId
-  }
-  deriving (Generic)
-instance NFData Goal
-
--- TODO: Is this a reasonable Eq instance?
-instance Eq Goal where
-  g1 == g2 = goalMeta g1 == goalMeta g2
-
--- | Components that are not changed during search. Components that do change
--- (local variables and let bindings) are stored in each 'SearchBranch'.
-data BaseComponents = BaseComponents
-  { hintFns :: [Component]
-  , hintDataTypes :: [Component]
-  , hintRecordTypes :: [Component]
-  , hintAxioms :: [Component]
-  -- ^ Excluding those producing Level
-  , hintLevel :: [Component]
-  -- ^ A definition in a where clause
-  , hintProjections :: [Component]
-  -- ^ Variables that are candidates for arguments to recursive calls
-  , hintThisFn :: Maybe Component
-  , hintLetVars :: [Open Component]
-  , hintRecVars :: Open [(Term, NoSubst Term Int)] -- ^ Variable terms and which argument they come from
-  , hintSplitVars :: Open [Term]
-  }
-  deriving (Generic)
-
-instance NFData BaseComponents
-
-type CompId = Int
-data Component = Component
-  { compId    :: CompId -- ^ Unique id for the component. Used for the cache.
-  , compName  :: Maybe Name -- ^ Used for keeping track of how many times a component has been used
-  , compPars  :: Nat -- ^ How many arguments should be dropped (e.g. constructor parameters)
-  , compTerm  :: Term
-  , compType  :: Type
-  , compRec   :: Bool -- ^ Is this a recursive call
-  , compMetas :: [MetaId]
-  , compCost  :: Cost
-  }
-  deriving (Eq, Generic)
-
-instance NFData Component
-
--- TODO: Is this reasonable?
-instance Ord Component where
-  compare = compare `on` compId
-
-data SearchStepResult
-  = ResultExpr Expr
-  | ResultClauses [A.Clause]
-  | OpenBranch SearchBranch
-  | NoSolution
-  deriving (Generic)
-instance NFData SearchStepResult
-
-
-data SearchOptions = SearchOptions
-  { searchBaseComponents :: BaseComponents
-  , searchHintMode :: HintMode
-  , searchTimeout :: MilliSeconds
-  , searchGenProjectionsLocal :: Bool
-  , searchGenProjectionsLet :: Bool
-  , searchGenProjectionsExternal :: Bool
-  , searchGenProjectionsRec :: Bool
-  , searchSpeculateProjections :: Bool
-  , searchTopMeta :: MetaId
-  , searchTopEnv :: TCEnv
-  , searchTopCheckpoint :: CheckpointId
-  , searchInteractionId :: InteractionId
-  , searchFnName :: Maybe QName
-  , searchCosts :: Costs
-  , searchStats :: IORef MimerStats
-  , searchRewrite :: Rewrite
-  , searchBuiltinFlat :: Maybe QName
-      -- Cache BUILTIN_FLAT for issue #7662 workaround
-  }
-
-type Cost = Int
-data Costs = Costs
-  { costLocal :: Cost
-  , costFn :: Cost
-  , costDataCon :: Cost
-  , costRecordCon :: Cost
-  , costSpeculateProj :: Cost
-  , costProj :: Cost
-  , costAxiom :: Cost
-  , costLet :: Cost
-  , costLevel :: Cost
-  , costSet :: Cost -- Should probably be replaced with multiple different costs
-  , costRecCall :: Cost
-  , costNewMeta :: Cost -- ^ Cost of a new meta-variable appearing in a non-implicit position
-  , costNewHiddenMeta :: Cost -- ^ Cost of a new meta-variable appearing in an implicit position
-  , costCompReuse :: Nat -> Cost -- ^ Cost of reusing a component @n@ times. Only counted when @n>1@.
-  }
-
-noCost :: Cost
-noCost = 0
-
-defaultCosts :: Costs
-defaultCosts = Costs
-  { costLocal = 3
-  , costFn = 10
-  , costDataCon = 3
-  , costRecordCon = 3
-  , costSpeculateProj = 20
-  , costProj = 3
-  , costAxiom = 10
-  , costLet = 5
-  , costLevel = 3
-  , costSet = 10
-  , costRecCall = 8
-  , costNewMeta = 10
-  , costNewHiddenMeta = 1
-  , costCompReuse = \uses -> 10 * (uses - 1) ^ 2
-  }
 
 ------------------------------------------------------------------------------
 -- * Helper functions
@@ -701,41 +540,6 @@ collectLHSVars ii = do
 
 declarationQnames :: A.Declaration -> [QName]
 declarationQnames dec = [ q | Scope.WithKind _ q <- A.declaredNames dec ]
-
-------------------------------------------------------------------------------
--- * Measure performance
-------------------------------------------------------------------------------
-data MimerStats = MimerStats
-  { statCompHit :: Nat -- ^ Could make use of an already generated component
-  , statCompGen :: Nat -- ^ Could use a generator for a component
-  , statCompRegen :: Nat -- ^ Had to regenerate the cache (new context)
-  , statCompNoRegen :: Nat -- ^ Did not have to regenerate the cache
-  , statMetasCreated :: Nat -- ^ Total number of meta-variables created explicitly (not through unification)
-  , statTypeEqChecks :: Nat -- ^ Number of times type equality is tested (with unification)
-  , statRefineSuccess :: Nat -- ^ Number of times a refinement has been successful
-  , statRefineFail :: Nat -- ^ Number of times a refinement has failed
-  } deriving (Show, Eq, Generic)
-instance NFData MimerStats
-
-emptyMimerStats :: MimerStats
-emptyMimerStats = MimerStats
-  { statCompHit = 0, statCompGen = 0, statCompRegen = 0 , statCompNoRegen = 0 , statMetasCreated = 0, statTypeEqChecks = 0, statRefineSuccess = 0 , statRefineFail = 0}
-
-incCompHit, incCompGen, incCompRegen, incCompNoRegen, incMetasCreated, incTypeEqChecks, incRefineSuccess, incRefineFail :: MimerStats -> MimerStats
-incCompHit       stats = stats {statCompHit       = succ $ statCompHit stats}
-incCompGen       stats = stats {statCompGen       = succ $ statCompGen stats}
-incCompRegen     stats = stats {statCompRegen     = succ $ statCompRegen stats}
-incCompNoRegen   stats = stats {statCompNoRegen   = succ $ statCompNoRegen stats}
-incMetasCreated  stats = stats {statMetasCreated  = succ $ statMetasCreated stats}
-incTypeEqChecks  stats = stats {statTypeEqChecks  = succ $ statTypeEqChecks stats}
-incRefineSuccess stats = stats {statRefineSuccess = succ $ statRefineSuccess stats}
-incRefineFail    stats = stats {statRefineFail    = succ $ statRefineFail stats}
-
-updateStat :: (MimerStats -> MimerStats) -> SM ()
-updateStat f = verboseS "mimer.stats" 10 $ do
-  ref <- asks searchStats
-  liftIO $ modifyIORef' ref f
-
 
 ------------------------------------------------------------------------------
 -- * Core algorithm
